dom/workers/test/serviceworkers/test_third_party_iframes.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<!DOCTYPE HTML>
<html>
<head>
  <meta http-equiv="Content-type" content="text/html;charset=UTF-8">
  <title>Bug 1152899 - Disallow the interception of third-party iframes using service workers when the third-party cookie preference is set</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">
<iframe></iframe>
</div>
<pre id="test"></pre>
<script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();
SimpleTest.requestLongerTimeout(2);

let index = 0;
function next() {
  info("Step " + index);
  if (index >= steps.length) {
    SimpleTest.finish();
    return;
  }
  try {
    let i = index++;
    steps[i]();
  } catch(ex) {
    ok(false, "Caught exception", ex);
  }
}

onload = next;

let iframe;
let basePath = "/tests/dom/workers/test/serviceworkers/thirdparty/";
let origin = window.location.protocol + "//" + window.location.host;
let thirdPartyOrigin = "https://example.com";

function testIframeLoaded() {
  ok(true, "Iframe loaded");
  iframe.removeEventListener("load", testIframeLoaded);
  let message = {
    source: "parent",
    href: origin + basePath + "iframe2.html"
  };
  iframe.contentWindow.postMessage(message.toSource(), "*");
}

function loadThirdPartyIframe() {
  let message = {
    source: "parent",
    href: thirdPartyOrigin + basePath + "iframe2.html"
  }
  iframe.contentWindow.postMessage(message.toSource(), "*");
}

function runTest(aExpectedResponses) {
  iframe = document.querySelector("iframe");
  iframe.src = thirdPartyOrigin + basePath + "register.html";
  let responsesIndex = 0;
  window.onmessage = function(e) {
    let status = e.data.status;
    let expected = aExpectedResponses[responsesIndex];
    if (status == expected.status) {
      ok(true, "Received expected " + expected.status);
      if (expected.next) {
        expected.next();
      }
    } else {
      ok(false, "Expected " + expected.status + " got " + status);
    }
    responsesIndex++;
  };
}

// Verify that we can register and intercept a 3rd party iframe with
// the given cookie policy.
function testShouldIntercept(policy, done) {
  SpecialPowers.pushPrefEnv({"set": [
      ["network.cookie.cookieBehavior", policy]
  ]}, function() {
    runTest([{
      status: "ok"
    }, {
      status: "registrationdone",
      next: function() {
        iframe.addEventListener("load", testIframeLoaded);
        iframe.src = origin + basePath + "iframe1.html";
      }
    }, {
      status: "networkresponse",
    }, {
      status: "worker-networkresponse",
      next: loadThirdPartyIframe
    }, {
      status: "swresponse",
    }, {
      status: "worker-swresponse",
      next: function() {
        iframe.src = thirdPartyOrigin + basePath + "unregister.html";
      }
    }, {
      status: "controlled",
    }, {
      status: "unregistrationdone",
      next: function() {
        window.onmessage = null;
        ok(true, "Test finished successfully");
        done();
      }
    }]);
  });
}

// Verify that we cannot register a service worker in a 3rd party
// iframe with the given cookie policy.
function testShouldNotRegister(policy, done) {
  SpecialPowers.pushPrefEnv({"set": [
      ["network.cookie.cookieBehavior", policy]
  ]}, function() {
    runTest([{
      status: "registrationfailed",
      next: function() {
        iframe.addEventListener("load", testIframeLoaded);
        iframe.src = origin + basePath + "iframe1.html";
      }
    }, {
      status: "networkresponse",
    }, {
      status: "worker-networkresponse",
      next: loadThirdPartyIframe
    }, {
      status: "networkresponse",
    }, {
      status: "worker-networkresponse",
      next: function() {
        window.onmessage = null;
        ok(true, "Test finished successfully");
        done();
      }
    }]);
  });
}

// Verify that if a service worker is already registered a 3rd
// party iframe will still not be intercepted with the given cookie
// policy.
function testShouldNotIntercept(policy, done) {
  SpecialPowers.pushPrefEnv({"set": [
      ["network.cookie.cookieBehavior", COOKIE_BEHAVIOR_ACCEPT]
  ]}, function() {
    runTest([{
      status: "ok"
    }, {
      status: "registrationdone",
      next: function() {
        iframe.addEventListener("load", testIframeLoaded);
        SpecialPowers.pushPrefEnv({"set": [
            ["network.cookie.cookieBehavior", policy],
          ]}, function() {
            iframe.src = origin + basePath + "iframe1.html";
          });
      }
    }, {
      status: "networkresponse",
    }, {
      status: "worker-networkresponse",
      next: loadThirdPartyIframe
    }, {
      status: "networkresponse",
    }, {
      status: "worker-networkresponse",
      next: function() {
        iframe.src = thirdPartyOrigin + basePath + "unregister.html";
      }
    }, {
      status: "uncontrolled",
    }, {
      status: "getregistrationfailed",
      next: function() {
        SpecialPowers.pushPrefEnv({"set": [
            ["network.cookie.cookieBehavior", COOKIE_BEHAVIOR_ACCEPT],
          ]}, function() {
            iframe.src = thirdPartyOrigin + basePath + "unregister.html";
          });
      }
    }, {
      status: "controlled",
    }, {
      status: "unregistrationdone",
      next: function() {
        window.onmessage = null;
        ok(true, "Test finished successfully");
        done();
      }
    }]);
  });
}

const COOKIE_BEHAVIOR_ACCEPT        = 0;
const COOKIE_BEHAVIOR_REJECTFOREIGN = 1;
const COOKIE_BEHAVIOR_REJECT        = 2;
const COOKIE_BEHAVIOR_LIMITFOREIGN  = 3;

let steps = [() => {
  SpecialPowers.pushPrefEnv({"set": [
    ["dom.serviceWorkers.exemptFromPerDomainMax", true],
    ["dom.serviceWorkers.enabled", true],
    ["dom.serviceWorkers.testing.enabled", true],
    ["browser.dom.window.dump.enabled", true],
    ["network.cookie.cookieBehavior", COOKIE_BEHAVIOR_ACCEPT]
  ]}, next);
}, () => {
  testShouldIntercept(COOKIE_BEHAVIOR_ACCEPT, next);
}, () => {
  testShouldNotRegister(COOKIE_BEHAVIOR_REJECTFOREIGN, next);
}, () => {
  testShouldNotIntercept(COOKIE_BEHAVIOR_REJECTFOREIGN, next);
}, () => {
  testShouldNotRegister(COOKIE_BEHAVIOR_REJECT, next);
}, () => {
  testShouldNotIntercept(COOKIE_BEHAVIOR_REJECT, next);
}, () => {
  testShouldNotRegister(COOKIE_BEHAVIOR_LIMITFOREIGN, next);
}, () => {
  testShouldNotIntercept(COOKIE_BEHAVIOR_LIMITFOREIGN, next);
}];

</script>
</pre>
</body>
</html>
